@@ -6,31 +6,30 @@
 
 // ReSharper disable CheckNamespace
 
-namespace Chapter.Net.WPF.Behaviors
+namespace Chapter.Net.WPF.Behaviors;
+
+/// <summary>
+///     Defines how the columns of a list view shall be resized by the <see cref="ColumnWidthBehavior" />.
+/// </summary>
+public enum ColumnResizeKind
 {
     /// <summary>
-    ///     Defines how the columns of a list view shall be resized by the <see cref="ColumnWidthBehavior" />.
+    ///     The column widths stays unchanged.
     /// </summary>
-    public enum ColumnResizeKind
-    {
-        /// <summary>
-        ///     The column widths stays unchanged.
-        /// </summary>
-        NoResize = -1,
+    NoResize = -1,
 
-        /// <summary>
-        ///     The column widths are defined by the owner control width.
-        /// </summary>
-        ByControl = -2,
+    /// <summary>
+    ///     The column widths are defined by the owner control width.
+    /// </summary>
+    ByControl = -2,
 
-        /// <summary>
-        ///     The column widths are defined by their shown content.
-        /// </summary>
-        ByContent = -3,
+    /// <summary>
+    ///     The column widths are defined by their shown content.
+    /// </summary>
+    ByContent = -3,
 
-        /// <summary>
-        ///     The column widths are calculated proportional by the owner control width.
-        /// </summary>
-        Proportional = -4
-    }
+    /// <summary>
+    ///     The column widths are calculated proportional by the owner control width.
+    /// </summary>
+    Proportional = -4
 }
